Cleaned up potential database integration
